toolkit/devtools/server/tests/mochitest/test_memory_allocations_02.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1067491 - Test aggregating allocation counts.
-->
<head>
  <meta charset="utf-8">
  <title>Memory monitoring actor test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="memory-helpers.js" type="application/javascript;version=1.8"></script>
<script>
window.onload = function() {
  SimpleTest.waitForExplicitFinish();

  Task.spawn(function* () {
    var { memory, client } = yield startServerAndGetSelectedTabMemory();
    yield memory.attach();

    yield memory.startRecordingAllocations();
    ok(true, "Can start recording allocations");

    // Allocate some objects.

    var allocs = [];
    (function allocator() {
      for (var i = 0; i < 10; i++) {
        allocs.push({});
      }
    }());

    var response = yield memory.getAllocations();

    yield memory.stopRecordingAllocations();
    ok(true, "Can stop recording allocations");

    // Find the index of our 10 allocations, and then assert that it is in the
    // `allocator` frame.

    var index = 0;
    var found = false;
    for (var count of response.counts) {
      if (count === 10) {
        found = true;
        break;
      }
      index++;
    }
    ok(found, "Should find the 10 allocations.");

    is(response.frames[index].functionDisplayName, "allocator",
       "Should have found the allocator frame.");

    yield memory.detach();
    destroyServerAndFinish(client);
  });
};
</script>
</pre>
</body>
</html>
